[New3D] Marker lifetimes, code cleanup, tf node icons
