+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/markers/RenderableArrow.ts
@@ -45,8 +45,8 @@ export class RenderableArrow extends RenderableMarker {
   shaftOutline: THREE.LineSegments | undefined;
   headOutline: THREE.LineSegments | undefined;
 
-  constructor(topic: string, marker: Marker, renderer: Renderer) {
-    super(topic, marker, renderer);
+  constructor(topic: string, marker: Marker, receiveTime: bigint | undefined, renderer: Renderer) {
+    super(topic, marker, receiveTime, renderer);
 
     // Shaft mesh
     const material = standardMaterial(marker.color, renderer.materialCache);
@@ -77,16 +77,16 @@ export class RenderableArrow extends RenderableMarker {
     this.headOutline.userData.picking = false;
     this.headMesh.add(this.headOutline);
 
-    this.update(marker);
+    this.update(marker, receiveTime);
   }
 
   override dispose(): void {
     releaseStandardMaterial(this.userData.marker.color, this._renderer.materialCache);
   }
 
-  override update(marker: Marker): void {
+  override update(marker: Marker, receiveTime: bigint | undefined): void {
     const prevMarker = this.userData.marker;
-    super.update(marker);
+    super.update(marker, receiveTime);
 
     if (!rgbaEqual(marker.color, prevMarker.color)) {
       releaseStandardMaterial(prevMarker.color, this._renderer.materialCache);
