+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/markers/RenderableCubeList.ts
@@ -19,8 +19,8 @@ export class RenderableCubeList extends RenderableMarker {
   mesh: DynamicInstancedMesh<THREE.BoxGeometry, THREE.Material>;
   // outline: THREE.LineSegments | undefined;
 
-  constructor(topic: string, marker: Marker, renderer: Renderer) {
-    super(topic, marker, renderer);
+  constructor(topic: string, marker: Marker, receiveTime: bigint | undefined, renderer: Renderer) {
+    super(topic, marker, receiveTime, renderer);
 
     // Cube instanced mesh
     const material = standardInstancedMaterial(marker, renderer.materialCache);
@@ -38,16 +38,16 @@ export class RenderableCubeList extends RenderableMarker {
     // this.outline.userData.picking = false;
     // this.add(this.outline);
 
-    this.update(marker);
+    this.update(marker, receiveTime);
   }
 
   override dispose(): void {
     releaseStandardInstancedMaterial(this.userData.marker, this._renderer.materialCache);
   }
 
-  override update(marker: Marker): void {
+  override update(marker: Marker, receiveTime: bigint | undefined): void {
     const prevMarker = this.userData.marker;
-    super.update(marker);
+    super.update(marker, receiveTime);
 
     if (markerHasTransparency(marker) !== markerHasTransparency(prevMarker)) {
       releaseStandardInstancedMaterial(prevMarker, this._renderer.materialCache);
