+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/markers/RenderableMarker.ts
@@ -4,6 +4,8 @@
 
 import * as THREE from "three";
 
+import { toNanoSec } from "@foxglove/rostime";
+
 import type { Renderer } from "../../Renderer";
 import { Marker, Pose, rosTimeToNanoSec } from "../../ros";
 import { getMarkerId } from "./markerId";
@@ -17,6 +19,8 @@ type MarkerUserData = {
   marker: Marker;
   pose: Pose;
   srcTime: bigint;
+  receiveTime: bigint | undefined;
+  expiresIn: bigint | undefined;
 };
 
 export class RenderableMarker extends THREE.Object3D {
@@ -24,10 +28,12 @@ export class RenderableMarker extends THREE.Object3D {
 
   protected _renderer: Renderer;
 
-  constructor(topic: string, marker: Marker, renderer: Renderer) {
+  constructor(topic: string, marker: Marker, receiveTime: bigint | undefined, renderer: Renderer) {
     super();
 
     this._renderer = renderer;
+
+    const hasLifetime = marker.lifetime.sec !== 0 || marker.lifetime.nsec !== 0;
 
     this.name = getMarkerId(topic, marker.ns, marker.id);
     this.userData = {
@@ -35,19 +41,23 @@ export class RenderableMarker extends THREE.Object3D {
       marker,
       pose: marker.pose,
       srcTime: rosTimeToNanoSec(marker.header.stamp),
+      receiveTime,
+      expiresIn: hasLifetime ? toNanoSec(marker.lifetime) : undefined,
     };
-
-    renderer.renderables.set(this.name, this);
   }
 
   dispose(): void {
-    this._renderer.renderables.delete(this.name);
+    //
   }
 
-  update(marker: Marker): void {
+  update(marker: Marker, receiveTime: bigint | undefined): void {
+    const hasLifetime = marker.lifetime.sec !== 0 || marker.lifetime.nsec !== 0;
+
     this.userData.marker = marker;
     this.userData.srcTime = rosTimeToNanoSec(marker.header.stamp);
     this.userData.pose = marker.pose;
+    this.userData.receiveTime = receiveTime;
+    this.userData.expiresIn = hasLifetime ? toNanoSec(marker.lifetime) : undefined;
   }
 
   // Convert sRGB values to linear
