+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/markers/RenderableSphere.ts
@@ -19,8 +19,8 @@ export class RenderableSphere extends RenderableMarker {
 
   mesh: THREE.Mesh<THREE.SphereGeometry, THREE.Material>;
 
-  constructor(topic: string, marker: Marker, renderer: Renderer) {
-    super(topic, marker, renderer);
+  constructor(topic: string, marker: Marker, receiveTime: bigint | undefined, renderer: Renderer) {
+    super(topic, marker, receiveTime, renderer);
 
     // Sphere mesh
     const material = standardMaterial(marker.color, renderer.materialCache);
@@ -29,16 +29,16 @@ export class RenderableSphere extends RenderableMarker {
     this.mesh.receiveShadow = true;
     this.add(this.mesh);
 
-    this.update(marker);
+    this.update(marker, receiveTime);
   }
 
   override dispose(): void {
     releaseStandardMaterial(this.userData.marker.color, this._renderer.materialCache);
   }
 
-  override update(marker: Marker): void {
+  override update(marker: Marker, receiveTime: bigint | undefined): void {
     const prevMarker = this.userData.marker;
-    super.update(marker);
+    super.update(marker, receiveTime);
 
     if (!rgbaEqual(marker.color, prevMarker.color)) {
       releaseStandardMaterial(prevMarker.color, this._renderer.materialCache);
