+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/markers/RenderableSphereList.ts
@@ -18,8 +18,8 @@ import {
 export class RenderableSphereList extends RenderableMarker {
   mesh: DynamicInstancedMesh<THREE.SphereGeometry, THREE.Material>;
 
-  constructor(topic: string, marker: Marker, renderer: Renderer) {
-    super(topic, marker, renderer);
+  constructor(topic: string, marker: Marker, receiveTime: bigint | undefined, renderer: Renderer) {
+    super(topic, marker, receiveTime, renderer);
 
     // Sphere instanced mesh
     const material = standardInstancedMaterial(marker, renderer.materialCache);
@@ -32,16 +32,16 @@ export class RenderableSphereList extends RenderableMarker {
     this.mesh.receiveShadow = true;
     this.add(this.mesh);
 
-    this.update(marker);
+    this.update(marker, receiveTime);
   }
 
   override dispose(): void {
     releaseStandardInstancedMaterial(this.userData.marker, this._renderer.materialCache);
   }
 
-  override update(marker: Marker): void {
+  override update(marker: Marker, receiveTime: bigint | undefined): void {
     const prevMarker = this.userData.marker;
-    super.update(marker);
+    super.update(marker, receiveTime);
 
     if (markerHasTransparency(marker) !== markerHasTransparency(prevMarker)) {
       releaseStandardInstancedMaterial(prevMarker, this._renderer.materialCache);
